Fixed slider click/drag coordinate errors in Gtk3.14
GdkEvent coordinates should be refer to GdkWindow and indeed
in Gtk3.14 that is the case so no extra translation is needed.
However, in Gtk3.12, the coordinates refer to the slider widget,
even though it has no GdkWindow associated with it.
TBC if there is a better way of handling this, but Gtk version
checking works well.